# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/fleets/{id}": {
      "get": {
        "operationId": "FleetsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Fleet id (uuid).",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The fleet.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/FleetDto"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "404": {
            "description": "Fleet not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch one fleet by id (platform path).",
        "tags": [
          "fleets"
        ]
      },
      "patch": {
        "operationId": "FleetsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Fleet id (uuid).",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFleetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated fleet.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/FleetDto"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "404": {
            "description": "Fleet not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a fleet (platform path).",
        "tags": [
          "fleets"
        ]
      },
      "delete": {
        "operationId": "FleetsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Fleet id (uuid).",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fleet deleted; data is null.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "nullable": true,
                      "example": null
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "404": {
            "description": "Fleet not found."
          },
          "409": {
            "description": "Fleet still referenced by buses/trips/bookings."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a fleet and its memberships (platform path, cascading).",
        "tags": [
          "fleets"
        ]
      }
    },
    "/fleets/{fleetId}/members": {
      "post": {
        "operationId": "MembersController_add",
        "parameters": [
          {
            "name": "fleetId",
            "required": true,
            "in": "path",
            "description": "Fleet selector (uuid) — a selector only; membership is verified server-side.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Membership created (defaults to ACTIVE).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 201
                    },
                    "data": {
                      "$ref": "#/components/schemas/FleetMemberDto"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "404": {
            "description": "Target user or role not found/inactive (404)."
          },
          "409": {
            "description": "The user is already a member of this fleet."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a user to the fleet with a role (members.manage required).",
        "tags": [
          "fleet-members"
        ]
      },
      "get": {
        "operationId": "MembersController_list",
        "parameters": [
          {
            "name": "fleetId",
            "required": true,
            "in": "path",
            "description": "Fleet selector (uuid) — a selector only; membership is verified server-side.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque nextCursor from the previous page (base64url of the last item id).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (default 20, max 100).",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cursor page of memberships (items + nextCursor).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/FleetMemberDto"
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the fleet memberships (cursor pagination).",
        "tags": [
          "fleet-members"
        ]
      }
    },
    "/fleets/{fleetId}/members/{memberId}": {
      "patch": {
        "operationId": "MembersController_update",
        "parameters": [
          {
            "name": "fleetId",
            "required": true,
            "in": "path",
            "description": "Fleet selector (uuid) — a selector only; membership is verified server-side.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "description": "Membership id (uuid).",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated membership.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/FleetMemberDto"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "404": {
            "description": "Membership not found in this fleet (cross-fleet ids are also 404)."
          },
          "409": {
            "description": "Role slug unknown/inactive for the resolution strategy."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change a member role/status (suspension revokes sessions immediately).",
        "tags": [
          "fleet-members"
        ]
      },
      "delete": {
        "operationId": "MembersController_remove",
        "parameters": [
          {
            "name": "fleetId",
            "required": true,
            "in": "path",
            "description": "Fleet selector (uuid) — a selector only; membership is verified server-side.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "description": "Membership id (uuid).",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Membership removed; data is null.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "nullable": true,
                      "example": null
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, revoked session, stale authVersion, or deactivated account."
          },
          "403": {
            "description": "Authenticated but lacking the required permission (or no ACTIVE membership)."
          },
          "404": {
            "description": "Membership not found in this fleet (cross-fleet ids are also 404)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a membership (target sessions invalidated immediately).",
        "tags": [
          "fleet-members"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bus Fleet API",
    "description": "Multi-tenant group bus transportation platform REST API.\n\n- **Authentication is owned by this API**: HS256 JWTs carry a custom `app_role` claim; refresh tokens rotate on every use and are stored hashed at rest.\n- **Every response is wrapped in a `{ statusCode, data }` envelope**; the `data` schema is shown inside each documented response.\n- **Fleet isolation is enforced by PostgreSQL RLS** as the final security boundary; cross-tenant lookups are indistinguishable from missing rows and return 404.\n- `fleetId` path parameters are **selectors, never proof of authorization** — access is verified against the caller's ACTIVE `FleetMember` row on every request.\n- Roles and permissions are **database-driven** (no hardcoded enums); only `super_admin` is a predefined system role. Platform routes (roles, permissions, users, fleets admin, audit-logs) are privileged: super admin only and audited.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "FleetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Fleet A"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "ownerId",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateFleetDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Fleet A",
            "minLength": 1,
            "maxLength": 255
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "FleetMemberDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "fleetId": {
            "type": "string",
            "format": "uuid"
          },
          "roleId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "SUSPENDED",
              "REVOKED"
            ],
            "example": "ACTIVE"
          },
          "joinedAt": {
            "format": "date-time",
            "type": "string"
          },
          "assignedBy": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "required": [
          "id",
          "userId",
          "fleetId",
          "roleId",
          "status",
          "joinedAt"
        ]
      },
      "AddMemberDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "Target user — must exist and be active (else 404)."
          },
          "roleSlug": {
            "type": "string",
            "example": "fleet-operator",
            "description": "Role slug within this fleet; roleId takes precedence when both are given."
          },
          "roleId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "SUSPENDED",
              "REVOKED"
            ],
            "default": "ACTIVE"
          }
        },
        "required": [
          "userId"
        ]
      },
      "UpdateMemberDto": {
        "type": "object",
        "properties": {
          "roleSlug": {
            "type": "string",
            "example": "fleet-operator"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "SUSPENDED",
              "REVOKED"
            ],
            "description": "Suspending/revoking takes effect immediately (sessions invalidated)."
          }
        }
      },
...
    }
  }
}